+++ src/ip_log.c
@@ -1,3 +1,13 @@
+/**
+ * @file ip_log.c
+ * @author Aryan Chopra
+ * @brief Logs the IP packet in IP log file.
+ *
+ * Opens the file at the location in append mode.
+ * Logs the IP address in decimal notation after converting it.
+ * Logs other necessary information in an IP packet.
+ */
+
 #include <arpa/inet.h>
 #include <errno.h>
 #include <inttypes.h>
@@ -17,7 +27,21 @@
 #define THIRD_OCTATE 0xFF00
 #define FOURTH_OCTATE 0xFF
 
+/**
+ * @brief Contains the file descirptor of the Log file opened.
+ */
+
 int ipLogFile; 
+
+/**
+ * @brief Opens the log file specified.
+ *
+ *
+ * Opens the log file at the specified location in append mode.
+ * In case of an error, it prints an explanatory error message on console and exits the process.
+ *
+ * @pre Log file exists at the location specified
+ */
 
 void openIpLog() {
   ipLogFile = open(LOG_LOCATION "ip_log.txt", O_WRONLY | O_APPEND);
@@ -27,9 +51,29 @@ void openIpLog() {
   }
 }
 
+/**
+ * @brief Writes the char array in the IP log file.
+ *
+ * @pre The file descriptor is valid.
+ */
+
 void writeIpLog(char *text) {
   write(ipLogFile, text, strlen(text));
 }
+
+/**
+ * @brief Converts the provided address from binary, network notation(Big Endian) 
+ * to the legible decimal notation.
+ *
+ *
+ * The function is static to avoid linking errors with the logIpAddress function in a different file.
+ * The function writes the formated string using snprintf to a buffer which is then logged to the log file.
+ * inet_atoa requires in_addr struct which contains a single field equivalent to uint32_t.
+ * The address provided is casted to the struct which is then used to covert the 
+ * IP address in commonl understandable format.
+ *
+ * @param[in] address 32 bit unsigned int which contains the IP address in binary, Big Endian notation.
+ */
 
 static void logIpAddress(uint32_t address) {
   char *text = calloc(SIZE, 1);
@@ -41,7 +85,19 @@ static void logIpAddress(uint32_t address) {
   free(text);
 }
 
-void logIpHeader(IpHeader *hdr, int incoming) {
+/**
+ * @brief Logs the IP Header in the Log file.
+ *
+ * Marks the IP Header as incoming or outgoing in the log file.
+ * Marking depends on the value of the flag provided.
+ * Logs the fields of the IP Header in the Log file.
+ * Calls necessary functions to convert the IP address in legible form.
+ *
+ * @param[in] hdr A struct formatted with apt name and sizes of fields in an IP Header.
+ * @param[in] incoming A flag which marks whether the header provided is incoming or outgoing.
+ */
+
+void logIpHeader(IpHeader *hdr, uint8_t incoming) {
   char *text = calloc(SIZE, 1);
 
   if (incoming) {
